+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Oct 15, 2107
 
 * :tada: Layer Support Polygon Data
+
+## Oct 9, 2017
+
+* :tada: Layer Support Point and Line Data